read stream to provide array arguments to consumer stream from array argument, this repalces FindFiles functionality
